requirements.txt
pandas==2.2.3
streamlit==1.44.1
plotly==5.24.1
requests 
python-dotenv